[Behat] Extract ProductAssociationsTrait from product pages
