+++ src/Sylius/Behat/Page/Admin/Product/SimpleProduct/UpdateSimpleProductPage.php
@@ -17,6 +17,7 @@ use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Session;
 use Sylius\Behat\Behaviour\ChecksCodeImmutability;
 use Sylius\Behat\Page\Admin\Crud\UpdatePage as BaseUpdatePage;
+use Sylius\Behat\Page\Admin\Product\Common\ProductAssociationsTrait;
 use Sylius\Behat\Page\Admin\Product\Common\ProductMediaTrait;
 use Sylius\Behat\Service\AutocompleteHelper;
 use Sylius\Behat\Service\DriverHelper;
@@ -30,6 +31,7 @@ use Symfony\Component\Routing\RouterInterface;
 class UpdateSimpleProductPage extends BaseUpdatePage implements UpdateSimpleProductPageInterface
 {
     use ChecksCodeImmutability;
+    use ProductAssociationsTrait;
     use ProductMediaTrait;
     use SimpleProductFormTrait;
 
@@ -324,6 +326,7 @@ class UpdateSimpleProductPage extends BaseUpdatePage implements UpdateSimpleProd
             ],
             $this->getDefinedFormElements(),
             $this->getDefinedProductMediaElements(),
+            $this->getDefinedProductAssociationsElements(),
         );
     }
